+++ src/Makefile.am
@@ -20,6 +20,7 @@ libantioch_la_SOURCES += parsing/src/ascii_parser.C
 libantioch_la_SOURCES += parsing/src/blottner_parsing.C
 libantioch_la_SOURCES += parsing/src/cea_mixture_parsing.C
 libantioch_la_SOURCES += parsing/src/cea_mixture_ascii_parsing.C
+libantioch_la_SOURCES += parsing/src/sutherland_parsing.C
 libantioch_la_SOURCES += parsing/src/transport_species_parsing.C
 
 # transport
